review: rename ec-error file to error
